stage-2_3DGS/depthsplat/inference_gradio/requirements.txt
# Requirements for DepthSplat Gradio Demo
# Most dependencies are shared with the main depthsplat package

gradio
numpy
torch
torchvision
Pillow
scipy
einops
omegaconf
hydra-core
